pyproject.toml
[tool.poetry]
name = "livechrg-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
python-dotenv = "^1.0.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.1.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.1.14"
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
mocker = "^1.1.1"
pandas = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
